+++ database/migrations/2021_06_21_100000_meteobridge_create_tables.php
@@ -22,6 +22,7 @@ class MeteobridgeCreateTables extends Migration
             $table->string('mb_version', 16)->nullable()->comment("Meteobridge software version");
             $table->float('latitude', 12, 8)->nullable();
             $table->float('longitude', 12, 8)->nullable();
+            $table->char('hash', 48)->nullable();
             $table->timestamps();
         });
 
@@ -34,10 +35,12 @@ class MeteobridgeCreateTables extends Migration
             $table->float('pressure', 7, 2)->nullable();
             $table->float('pressure_sea', 7, 2)->nullable();
             $table->float('wind', 6, 2)->nullable();
-            $table->float('wind_avg', 6, 2)->nullable();
+            $table->float('wind_avg', 6, 2)->nullable()->comment("Average wind since last report");
+            $table->float('wind_min', 6, 2)->nullable()->comment("Minimum wind since last report");
+            $table->float('wind_max', 6, 2)->nullable()->comment("Maximum wind (gust) since last report");
             $table->integer('direction')->nullable();
             $table->float('rain_rate', 6, 2)->nullable();
-            $table->float('rain_total', 6, 2)->nullable();
+            $table->float('rain_total', 6, 2)->nullable()->comment("Total precipitation for current day");
             $table->smallInteger('uv_index')->nullable();
             $table->integer('radiation')->nullable();
             $table->unique(['timestamp', 'station_id'], 'meteobridge_observations__station_timestamp_index');