+++ src/Console/StationHttpTemplate.php
@@ -13,7 +13,8 @@ class StationHttpTemplate extends Command
      * @var string
      */
     protected $signature = 'meteobridge:http-template
-                           { id : Station ID }';
+                           { id : Station ID }
+                           { --i|interval=10 : Intervals (in minutes) between expected observations }';
 
     /**
      * Map between db/model property and meteobridge template equivalent.
@@ -34,7 +35,6 @@ class StationHttpTemplate extends Command
         'pressure' => 'thb0press-act.2',
         'pressure_sea' => 'thb0seapress-act.2',
         'wind' => 'wind0wind-act.2',
-        'wind_avg' => 'wind0avgwind-act.2',
         'direction' => 'wind0dir-act.0',
         'rain_rate' => 'rain0rate-act.2',
         'rain_total' => 'rain0total-daysum.2',
@@ -71,12 +71,45 @@ class StationHttpTemplate extends Command
             $this->error("Station not found: " . $this->argument('id'));
             return 1;
         }
-        $base = route('meteobridge.observe', ['station' => $station->id]);
+        if (!$this->generateIntervalMacros()) {
+            return 1;
+        }
+        $base = route('meteobridge.observation', ['station' => $station->id, 'hash' => $station->hash ?: 'none']);
         $args = ['timestamp=[YYYY]-[MM]-[DD]T[hh]:[mm]:[ss]'];
         foreach ($this->argMap as $arg => $template) {
-            $args[] = sprintf("#if#{*[%s:-9999]!=-9999*}#then#&%s=[%s]#else##fi#", $template, $arg, $template);
+            $args[] = $this->macroIfThenElse(
+                sprintf("{*[%s:-9999]!=-9999*}", $template),
+                sprintf("&%s=[%s]", $arg, $template),
+                ''
+            );
         }
         $this->line($base . '?' . implode("", $args));
         return 0;
+    }
+
+    /**
+     * Generate the macros for min, max and avg values between reports.
+     *
+     * @return bool
+     */
+    protected function generateIntervalMacros()
+    {
+        $interval = intval($this->option('interval'), 10);
+        if ($interval < 1 || $interval > 60) {
+            $this->error("The interval can only be between 1 and 60 minutes");
+            return false;
+        }
+        $this->argMap['wind_max'] = "wind0wind-max{$interval}.2";
+        $this->argMap['wind_min'] = "wind0wind-min{$interval}.2";
+        $this->argMap['wind_avg'] = "wind0wind-avg{$interval}.2";
+        return true;
+    }
+
+    /**
+     * Wrapper around the if/then/else syntax in Metrobridge templates.
+     */
+    protected function macroIfThenElse($test, $thenVal, $elseVal = '')
+    {
+        return sprintf("#if#%s#then#%s#else#%s#fi#", $test, $thenVal, $elseVal);
     }
 }