+++ src/MeteobridgeServiceProvider.php
@@ -9,6 +9,7 @@ use TromsFylkestrafikk\Meteobridge\Console\StationAdd;
 use TromsFylkestrafikk\Meteobridge\Console\StationDelete;
 use TromsFylkestrafikk\Meteobridge\Console\StationHttpTemplate;
 use TromsFylkestrafikk\Meteobridge\Console\StationList;
+use TromsFylkestrafikk\Meteobridge\Console\StationRehash;
 use TromsFylkestrafikk\Meteobridge\Console\StationSetParam;
 
 class MeteobridgeServiceProvider extends ServiceProvider
@@ -50,6 +51,7 @@ class MeteobridgeServiceProvider extends ServiceProvider
                 StationDelete::class,
                 StationHttpTemplate::class,
                 StationList::class,
+                StationRehash::class,
                 StationSetParam::class,
             ]);
         }